Rename folder (remove "ariakit")
